updated adc handling and task prio TODO test the new adc 3 channel conversion stuff

+++ IoTM_/STM32CubeIDE/App/adc/adcHandler.h
@@ -24,7 +24,7 @@ public:
 	void adcConcCpltCallback(ADC_HandleTypeDef *hadc);
 
 	osMessageQueueId_t getQueue();
-	uint32_t* getBuffer();
+	uint16_t* getBuffer();
 
 	// Task Loop:
 	void run();
